add deprecated warning checks, rename tests
