+++ tests/integration_tests/spark/sdk/cluster/test_cluster_deprecated.py
@@ -14,20 +14,21 @@ from aztk.utils import constants
 from aztk_cli import config
 from tests.integration_tests.spark.sdk.get_client import get_spark_client, get_test_suffix
 
+
 base_cluster_id = get_test_suffix("cluster")
 spark_client = get_spark_client()
 
 
 def clean_up_cluster(cluster_id):
     try:
-        spark_client.cluster.delete(id=cluster_id)
+        spark_client.delete_cluster(cluster_id=cluster_id)
     except (BatchErrorException, AztkError):
         # pass in the event that the cluster does not exist
         pass
 
 
 def ensure_spark_master(cluster_id):
-    results = spark_client.cluster.run(cluster_id,
+    results = spark_client.cluster_run(cluster_id,
                 "if $AZTK_IS_MASTER ; then $SPARK_HOME/sbin/spark-daemon.sh status org.apache.spark.deploy.master.Master 1 ;" \
                 " else echo AZTK_IS_MASTER is false ; fi")
     for _, result in results:
@@ -38,7 +39,7 @@ def ensure_spark_master(cluster_id):
 
 
 def ensure_spark_worker(cluster_id):
-    results = spark_client.cluster.run(cluster_id,
+    results = spark_client.cluster_run(cluster_id,
                 "if $AZTK_IS_WORKER ; then $SPARK_HOME/sbin/spark-daemon.sh status org.apache.spark.deploy.worker.Worker 1 ;" \
                 " else echo AZTK_IS_WORKER is false ; fi")
     for _, result in results:
@@ -55,12 +56,10 @@ def ensure_spark_processes(cluster_id):
 def wait_for_all_nodes(cluster_id, nodes):
     while True:
         for node in nodes:
-            if node.state in [batch_models.ComputeNodeState.unusable, batch_models.ComputeNodeState.start_task_failed]:
-                raise AztkError("Node {} in failed state.".format(node.id))
             if node.state not in [batch_models.ComputeNodeState.idle, batch_models.ComputeNodeState.running]:
                 break
         else:
-            nodes = spark_client.cluster.get(cluster_id).nodes
+            nodes = spark_client.get_cluster(cluster_id).nodes
             continue
         break
 
@@ -79,7 +78,42 @@ def test_create_cluster():
         toolkit=aztk.spark.models.SparkToolkit(version="2.3.0"),
         spark_configuration=None)
     try:
-        cluster = spark_client.cluster.create(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            cluster = spark_client.create_cluster(cluster_configuration, wait=True)
+
+        assert cluster.pool is not None
+        assert cluster.nodes is not None
+        assert cluster.id == cluster_configuration.cluster_id
+        assert cluster.vm_size == "standard_f2"
+        assert cluster.current_dedicated_nodes == 2
+        assert cluster.gpu_enabled is False
+        assert cluster.master_node_id is not None
+        assert cluster.current_low_pri_nodes == 0
+
+    except (AztkError, BatchErrorException) as e:
+        assert False
+
+    finally:
+        clean_up_cluster(cluster_configuration.cluster_id)
+
+
+def test_get_cluster():
+    test_id = "test-get-"
+    cluster_configuration = aztk.spark.models.ClusterConfiguration(
+        cluster_id=test_id + base_cluster_id,
+        vm_count=2,
+        vm_low_pri_count=0,
+        vm_size="standard_f2",
+        subnet_id=None,
+        custom_scripts=None,
+        file_shares=None,
+        toolkit=aztk.spark.models.SparkToolkit(version="2.3.0"),
+        spark_configuration=None)
+    try:
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            cluster = spark_client.get_cluster(cluster_id=cluster_configuration.cluster_id)
 
         assert cluster.pool is not None
         assert cluster.nodes is not None
@@ -110,8 +144,10 @@ def test_list_clusters():
         toolkit=aztk.spark.models.SparkToolkit(version="2.3.0"),
         spark_configuration=None)
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
-        clusters = spark_client.cluster.list()
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            clusters = spark_client.list_clusters()
 
         assert cluster_configuration.cluster_id in [cluster.id for cluster in clusters]
 
@@ -135,9 +171,12 @@ def test_get_remote_login_settings():
         toolkit=aztk.spark.models.SparkToolkit(version="2.3.0"),
         spark_configuration=None)
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
-        cluster = spark_client.cluster.get(id=cluster_configuration.cluster_id)
-        rls = spark_client.cluster.get_remote_login_settings(id=cluster.id, node_id=cluster.master_node_id)
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            cluster = spark_client.get_cluster(cluster_id=cluster_configuration.cluster_id)
+        with pytest.warns(DeprecationWarning):
+            rls = spark_client.get_remote_login_settings(cluster_id=cluster.id, node_id=cluster.master_node_id)
 
         assert rls.ip_address is not None
         assert rls.port is not None
@@ -178,10 +217,13 @@ def test_submit():
         executor_cores=None,
         max_retry_count=None)
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
 
-        spark_client.cluster.submit(
-            id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            spark_client.submit(
+                cluster_id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
+        
         assert True
 
     except (AztkError, BatchErrorException):
@@ -219,15 +261,18 @@ def test_get_application_log():
         executor_cores=None,
         max_retry_count=None)
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
 
-        spark_client.cluster.submit(
-            id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
-        application_log = spark_client.cluster.get_application_log(
-            id=cluster_configuration.cluster_id,
-            application_name=application_configuration.name,
-            tail=False,
-            current_bytes=0)
+        with pytest.warns(DeprecationWarning):
+            spark_client.submit(
+                cluster_id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            application_log = spark_client.get_application_log(
+                cluster_id=cluster_configuration.cluster_id,
+                application_name=application_configuration.name,
+                tail=False,
+                current_bytes=0)
 
         assert application_log.exit_code == 0
         assert application_log.name == application_configuration.name == "pipy100"
@@ -280,12 +325,14 @@ def test_get_application_status_complete():
         executor_cores=None,
         max_retry_count=None)
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
-
-        spark_client.cluster.submit(
-            id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
-        status = spark_client.cluster.get_application_status(
-            id=cluster_configuration.cluster_id, application_name=application_configuration.name)
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            spark_client.submit(
+                cluster_id=cluster_configuration.cluster_id, application=application_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            status = spark_client.get_application_status(
+                cluster_id=cluster_configuration.cluster_id, app_name=application_configuration.name)
 
         assert status == "completed"
 
@@ -310,8 +357,9 @@ def test_delete_cluster():
         spark_configuration=None)
 
     try:
-        spark_client.cluster.create(cluster_configuration, wait=True)
-        success = spark_client.cluster.delete(id=cluster_configuration.cluster_id)
+        with pytest.warns(DeprecationWarning):
+            spark_client.create_cluster(cluster_configuration, wait=True)
+            success = spark_client.delete_cluster(cluster_id=cluster_configuration.cluster_id)
 
         assert success is True
 
@@ -336,9 +384,12 @@ def test_spark_processes_up():
         spark_configuration=None)
 
     try:
-        cluster = spark_client.cluster.create(cluster_configuration, wait=True)
-        wait_for_all_nodes(cluster.id, cluster.nodes)
-        success = spark_client.cluster.delete(id=cluster_configuration.cluster_id)
+        with pytest.warns(DeprecationWarning):
+            cluster = spark_client.create_cluster(cluster_configuration, wait=True)
+            wait_for_all_nodes(cluster.id, cluster.nodes)
+
+        with pytest.warns(DeprecationWarning):
+            success = spark_client.delete_cluster(cluster_id=cluster_configuration.cluster_id)
 
         assert success is True
 
@@ -366,10 +417,15 @@ def test_debug_tool():
         "spark/logs", "spark/wd"
     ]
     try:
-        cluster = spark_client.cluster.create(cluster_configuration, wait=True)
+        with pytest.warns(DeprecationWarning):
+            cluster = spark_client.create_cluster(cluster_configuration, wait=True)
+
         nodes = [node for node in cluster.nodes]
         wait_for_all_nodes(cluster.id, nodes)
-        cluster_output = spark_client.cluster.diagnostics(id=cluster.id)
+
+        with pytest.warns(DeprecationWarning):
+            cluster_output = spark_client.run_cluster_diagnostics(cluster_id=cluster.id)
+
         for node_output in cluster_output:
             node_output.output.seek(0)    # tempfile requires seek 0 before reading
             debug_zip = ZipFile(node_output.output)
